Add only missing method test cases  - Look for method invocations instead of method name in the test case title  - add CONTRIBUTING.md  - fix typos and syntax errors #closes #100
