+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "scuri",
-  "version": "1.0.3",
+  "version": "1.0.4",
   "description": "A spec generator schematic - Spec Create Update Read (class - component, service, directive and dependencies) Incorporate (them in the result)",
   "scripts": {
     "s": "schematics",
